+++ server/mongodb/models/nonHourlyShiftDetails.ts
